# docs/openapi/openapi-v1.yaml
openapi: 3.0.3
info:
  title: Fitness Tracker API
  version: 1.0.0
  description: Initial OpenAPI v1 for MVP (users/auth, sessions, goals).
servers:
  - url: https://api.example.com/v1
paths:
...
  /routines:
    get:
      summary: List routines
      tags: [Routines]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: List
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Routine' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Create routine
      tags: [Routines]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RoutineCreate' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Routine' }}}}
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /exercises:
    get:
      summary: List exercises
      tags: [Exercises]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: List, content: { application/json: { schema: { type: array, items: { $ref: '#/components/schemas/Exercise' }}}}}
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      summary: Create exercise
      tags: [Exercises]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ExerciseCreate' }
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Exercise' }}}}
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    BadRequest:
      description: Bad Request
    Unauthorized:
      description: Unauthorized
...
  schemas:
...
    Routine:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
    RoutineCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
...
    Exercise:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        sets: { type: integer, nullable: true }
        reps: { type: integer, nullable: true }
        weight: { type: number, format: float, nullable: true }
        equipment: { type: string, nullable: true }
        duration: { type: string, format: time, nullable: true }
        distance: { type: number, format: float, nullable: true }
        stretch_type: { type: string, nullable: true }
        is_strength: { type: boolean, nullable: true }
        is_cardio: { type: boolean, nullable: true }
        is_core: { type: boolean, nullable: true }
        is_plyo: { type: boolean, nullable: true }
        is_flexibility: { type: boolean, nullable: true }
    ExerciseCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        sets: { type: integer, nullable: true }
        reps: { type: integer, nullable: true }
        weight: { type: number, format: float, nullable: true }
        equipment: { type: string, nullable: true }
        duration: { type: string, format: time, nullable: true }
        distance: { type: number, format: float, nullable: true }
        stretch_type: { type: string, nullable: true }
        is_strength: { type: boolean, nullable: true }
        is_cardio: { type: boolean, nullable: true }
        is_core: { type: boolean, nullable: true }
        is_plyo: { type: boolean, nullable: true }
        is_flexibility: { type: boolean, nullable: true }
...
security:
  - bearerAuth: []